[TEMP] Stop gold path generation when iteration limit reached
For the two melting point tasks, gold path generation checks the
temperature repeatedly until the state changes or an iteration limit is
reached. If the state does not change within the allotted number of
iterations, then we need to check that the temperature threshold has
been crossed. If it has not, then path generation should stop instead of